Feat: isolate proposal timer so it doesn't rerender so much
